+++ pakyow-support/lib/pakyow/support/definable.rb
@@ -1,242 +1,340 @@
 # frozen_string_literal: true
 
+require "pakyow/support/class_state"
 require "pakyow/support/deep_dup"
-require "pakyow/support/inflector"
+require "pakyow/support/deprecatable"
+require "pakyow/support/extension"
+require "pakyow/support/isolable"
 require "pakyow/support/makeable"
+
+require "pakyow/support/definable/registry"
+require "pakyow/support/definable/state"
 
 module Pakyow
   module Support
-    # Provides control over how state is defined on an object, and how state is
-    # shared across object instances and subclasses.
-    #
-    # You define the type of state provided by an object, along with any global
-    # state for that object type. When an instance is created or the definable
-    # object is subclassed, the new object inherits the global state and can be
-    # extended with its own state. Definable objects' `initialize` method should
-    # always call super with the block to ensure that state is inherited correctly.
-    #
-    # Once `defined!` is called on an instance, consider freezing the object so
-    # that it cannot be extended later.
-    #
+    # Define state for an object.
     #
     # @example
-    #   class SomeDefinableObject
-    #     include Support::Definable
+    #   class Controller
+    #     ...
+    #   end
     #
-    #     def initialize(some_arg, &block)
-    #       super()
-
-    #       # Do something with some_arg, etc.
+    #   class Application
+    #     include Pakyow::Support::Definable
     #
-    #       defined!(&block)
+    #     # Create a definable subclass of `Controller`.
+    #     #
+    #     definable :controller, Controller do
+    #       # Optionally extend the subclass with custom behavior.
     #     end
     #   end
     #
-    # @api private
+    #   # Subclass `Application` to work within a unique namespace.
+    #   #
+    #   class MyApplication < Application
+    #     controller :foo do
+    #       # Define custom behavior for the `foo` controller.
+    #     end
+    #   end
+    #
+    #   # Access the `foo` controller on the class.
+    #   #
+    #   MyApplication.controllers.foo.new(...)
+    #
+    #   # State is also accessible on the instance.
+    #   #
+    #   MyApplication.new.controllers.foo.new(...)
+    #
+    #   # State is defined within a logical namespace.
+    #   #
+    #   MyApplication.controllers.foo.class
+    #   => MyApplication::Controllers::Foo
+    #
+    #   # Additional state can be defined through the initializer.
+    #   #
+    #   MyApplication.new(...) do
+    #     controller :bar
+    #   end
+    #
+    #   # Keyword arguments are automatically set as class-level instance variables.
+    #   #
+    #   MyApplication.new(...) do
+    #     controller :baz, path: "/baz"
+    #   end
+    #
+    #   MyApplication.controllers.foo.instance_variable_get(:@path)
+    #   => "/baz"
+    #
+    # = State Lookup
+    #
+    # Definable state can be looked up by name.
+    #
+    # @example
+    #   class Controller
+    #     ...
+    #   end
+    #
+    #   class Application
+    #     include Pakyow::Support::Definable
+    #
+    #     definable :controller, Controller
+    #   end
+    #
+    #   class MyApplication < Application
+    #     controller :foo
+    #   end
+    #
+    #   MyApplication.controller.foo
+    #   => MyApplication::Controllers::Foo
+    #
+    # = Custom Lookups
+    #
+    # Definable state can be registered with a lookup function for custom lookups.
+    #
+    # @example
+    #   class Controller
+    #     attr_reader :path
+    #
+    #     def initialize(app, path)
+    #       @path = path
+    #     end
+    #
+    #     ...
+    #   end
+    #
+    #   class Application
+    #     include Pakyow::Support::Definable
+    #
+    #     definable :controller, Controller, lookup: -> (app, controller, path) {
+    #       controller.new(app, path)
+    #     }
+    #   end
+    #
+    #   class MyApplication < Application
+    #     controller :foo
+    #   end
+    #
+    #   MyApplication.controller.foo("/").path
+    #   => "/"
+    #
+    # = Argument Builder
+    #
+    # Definable state can be defined with custom arguments by specifying an argument builder.
+    #
+    # @example
+    #   class Controller
+    #     ...
+    #   end
+    #
+    #   class Application
+    #     include Pakyow::Support::Definable
+    #
+    #     definable :controller, Controller, builder: -> (name, path = "/") {
+    #       return name, path: path
+    #     }
+    #   end
+    #
+    #   class MyApplication < Application
+    #     controller :foo, "/foo"
+    #   end
+    #
+    #   MyApplication.controller.foo.path
+    #   => "/foo"
+    #
+    # = Definable Context
+    #
+    # By default, isolated objects are defined within the definable object's namespace. This
+    # behavior can be changed by specifying a different context using `context`.
+    #
+    # @example
+    #   class Controller
+    #     ...
+    #   end
+    #
+    #   class Application
+    #     include Pakyow::Support::Definable
+    #
+    #     definable :controller, Controller, context: Object
+    #   end
+    #
+    #   class MyApplication < Application
+    #     controller :foo, "/foo"
+    #   end
+    #
+    #   MyApplication.controller.foo
+    #   => Foo
+    #
+    # = Prioritized State
+    #
+    # By default, defined state is exposed in the order it's defined. In cases where order matters,
+    # set the `priority` keyword argument to `:high` or `:low` (defaults to `:default`).
+    #
+    # @example
+    #   class Controller
+    #     ...
+    #   end
+    #
+    #   class Application
+    #     include Pakyow::Support::Definable
+    #
+    #     definable :controller, Controller
+    #   end
+    #
+    #   class MyApplication < Application
+    #     controller :foo, priority: :low
+    #
+    #     controller :bar, priority: :high
+    #
+    #     controller :baz
+    #   end
+    #
+    #   MyApplication.controllers.each.map { |controller|
+    #     controller.class
+    #   }
+    #
+    #   => [
+    #        MyApplication::Controllers::Bar,
+    #        MyApplication::Controllers::Baz,
+    #        MyApplication::Controllers::Foo
+    #      ]
     #
     module Definable
+      extend Extension
+
+      extend_dependency ClassState
+      include_dependency Isolable
+
       using DeepDup
 
-      def self.included(base)
-        base.include CommonMethods
-        base.extend ClassMethods, CommonMethods
-        base.prepend Initializer
-        base.extend Support::Makeable
-        base.instance_variable_set(:@__state, {})
+      apply_extension do
+        class_state :__definable_registries, default: {}, inheritable: true
       end
 
-      # @api private
-      def defined!(&block)
-        # set instance level state
-        self.instance_eval(&block) if block_given?
+      class_methods do
+        extend Deprecatable
 
-        # merge global state
-        @__state.each do |name, state|
-          state.instances.concat(self.class.__state[name].instances)
-        end
+        # Register a type of state that can be defined by `name`.
+        #
+        # @param state_name [Symbol] the name of the definable state
+        # @param definable_object [Object] the object to be defined
+        # @param builder [Proc] an optional argument builder
+        # @param lookup [Proc] an optional lookup function
+        # @param context [Object] the context to define state in
+        #
+        def definable(state_name, definable_object, builder: nil, lookup: nil, context: isolable_context, &block)
+          isolated_object = isolate definable_object, context: context do
+            DEFINABLE_OBJECT_EXTENSIONS.each do |extension|
+              unless ancestors.include?(extension)
+                include extension
+              end
+            end
 
-        # merge inherited state
-        if inherited = self.class.__inherited_state
-          @__state.each do |name, state|
-            instances = state.instances
-            instances.concat(inherited[name].instances) if inherited[name]
+            class_eval(&block) if block_given?
           end
+
+          namespace = if context.equal?(isolable_context)
+            [Support.inflector.pluralize(state_name.to_s).to_sym]
+          else
+            []
+          end
+
+          state_registry = Registry.new(
+            state_name,
+            isolated_object,
+            parent: self,
+            builder: builder,
+            lookup: lookup,
+            namespace: namespace
+          )
+
+          isolated_object.instance_variable_set(:@__defined_state, state_registry)
+          __definable_registries[state_name.to_sym] = state_registry
+
+          code = <<~CODE
+            def #{state_name}(*args, **kwargs, &block)
+              __definable_registries[#{state_name.to_sym.inspect}].define(*args, **kwargs, &block)
+            end
+
+            def #{Support.inflector.pluralize(state_name)}(*args)
+              registry = __definable_registries[#{state_name.to_sym.inspect}]
+
+              if args.any?
+                registry.find(*args)
+              else
+                registry
+              end
+            end
+          CODE
+
+          class_eval(code); singleton_class.class_eval(code)
         end
-      end
 
-      module ClassMethods
-        attr_reader :__state, :__inherited_state
+        # Define the object.
+        #
+        def define(&block)
+          class_eval(&block)
+        end
 
+        def state(type)
+          __definable_registries[type.to_sym]
+        end
+        deprecate :state, solution: "all the cooresponding method for `type'"
+
+        # @api private
+        DEFINABLE_OBJECT_EXTENSIONS = [
+          Makeable, State
+        ].freeze
+
+        # @api private
         def inherited(subclass)
           super
 
-          subclass.instance_variable_set(:@__inherited_state, @__state.deep_dup)
-          subclass.instance_variable_set(:@__state, @__state.each_with_object({}) { |(name, state_instance), state|
-            state[name] = State.new(name, state_instance.object)
-          })
-        end
-
-        # Register a type of state that can be defined.
-        #
-        # @param object
-        #   Can be a class or instance, but must respond to :make. The `make`
-        #   method should return the object to be "made" and accept a block
-        #   that should be evaluated in the context of the object.
-        #
-        #     class Person
-        #       # ...
-        #
-        #       def make(name, dob, &block)
-        #         person = self.class.new(name, dob)
-        #
-        #         person.instance_eval(&block)
-        #         person
-        #       end
-        #
-        #       def befriend(person)
-        #         friends << person
-        #       end
-        #     end
-        #
-        #     class Application
-        #       include Pakyow::Support::Definable
-        #
-        #       stateful :person, Person
-        #     end
-        #
-        #     john = Application.person 'John', Date.new(1988, 8, 13) do
-        #     end
-        #
-        #     Application.person 'Sofie', Date.new(2015, 9, 6) do
-        #       befriend(john)
-        #     end
-        #
-        def stateful(name, object, &stateful_block)
-          name = name.to_sym
-          @__state[name] = State.new(name, object)
-          plural_name = Support.inflector.pluralize(name.to_s).to_sym
-
-          within = if object_name
-            ObjectNamespace.new(*object_name.namespace.parts.dup.concat([plural_name]))
-          else
-            self
-          end
-
-          method_body = Proc.new do |*args, priority: :default, **opts, &block|
-            return @__state[name] if block.nil?
-
-            stateful_block.call(args, opts) if stateful_block
-            object.make(*args, within: within, **opts, &block).tap do |state|
-              @__state[name].register(state, priority: priority)
+          subclass.__definable_registries.each_value do |registry|
+            if registry.object.name&.start_with?(name)
+              registry.rebase(subclass.isolate(registry.object))
             end
-          end
 
-          define_method name, &method_body
-          define_singleton_method name, &method_body
-        end
-
-        # Define state for the object.
-        #
-        def define(&block)
-          class_exec(&block)
-        end
-      end
-
-      module CommonMethods
-        # Returns registered state instances. If +type+ is passed, returns state of that type.
-        #
-        def state(type = nil)
-          if instance_variable_defined?(:@__state)
-            return @__state if type.nil?
-
-            if @__state && @__state.key?(type)
-              @__state[type].instances
-            else
-              []
-            end
-          else
-            {}
+            registry.reparent(subclass)
           end
         end
       end
 
-      module Initializer
-        def initialize(*)
-          # Create mutable state for this instance based on global.
+      prepend_methods do
+        # @api private
+        def initialize(*, &block)
+          @__defined = false
+          @__definable_registries = self.class.__definable_registries.deep_dup
+          @__definable_registries.each_value do |registry|
+            registry.reparent(self)
+          end
+
+          # Call super first. Any state defined here will take precedence over state in `block`.
           #
-          @__state = self.class.__state.each_with_object({}) { |(name, global_state), state|
-            state[name] = State.new(name, global_state.object)
-          }
-
           super
-        end
-      end
-    end
 
-    # Contains state for a definable class or instance.
-    #
-    # @api private
-    class State
-      using DeepDup
-
-      attr_reader :name, :object, :instances, :priorities
-
-      PRIORITIES = { default: 0, high: 1, low: -1 }.freeze
-
-      def initialize(name, object)
-        @name = name.to_sym
-        @object = object
-        @instances = []
-        @priorities = {}
-      end
-
-      def initialize_copy(original)
-        super
-        @instances = original.instances.deep_dup
-      end
-
-      # TODO: we handle both instances and classes, so reconsider the variable naming
-      def <<(instance)
-        register(instance)
-      end
-
-      # TODO: we handle both instances and classes, so reconsider the variable naming
-      def register(instance, priority: :default)
-        unless priority.is_a?(Integer)
-          priority = PRIORITIES.fetch(priority) {
-            raise ArgumentError, "unknown priority `#{priority}'"
-          }
-        end
-
-        unless instance.is_a?(Module)
-          enforce_registration!(instance)
-        end
-
-        unless instances.include?(instance)
-          instances << instance
-        end
-
-        priorities[instance] = priority
-        reprioritize!
-      end
-
-      def enforce_registration!(instance)
-        ancestors = if instance.respond_to?(:new)
-          instance.ancestors
-        else
-          instance.class.ancestors
-        end
-
-        unless ancestors.include?(@object)
-          raise ArgumentError, "expected instance of '#{@object}'"
+          define!(&block)
         end
       end
 
-      def reprioritize!
-        @instances.sort! { |a, b|
-          (@priorities[b] || 0) <=> (@priorities[a] || 0)
-        }
+      # Automatically called at the end of initializing the definable object. Call this method
+      # yourself to control when the definable block is called. It won't be called again.
+      #
+      def define!(&block)
+        unless @__defined
+          @__defined = true
+
+          if block_given?
+            instance_eval(&block)
+          end
+        end
       end
+
+      def state(type)
+        self.class.state(type)
+      end
+
+      # @api private
+      attr_reader :__definable_registries
     end
   end
 end